Fixed review findings part 2
