+++ exasol/ds/sandbox/cli/commands/setup_vm_bucket_waf.py
@@ -7,7 +7,7 @@ from exasol.ds.sandbox.cli.options.logging import logging_options
 from exasol.ds.sandbox.lib.aws_access.aws_access import AwsAccess
 from exasol.ds.sandbox.lib.config import default_config_object
 from exasol.ds.sandbox.lib.logging import set_log_level
-from exasol.ds.sandbox.lib.cloudformation.s3_buckets import VmBucket
+from exasol.ds.sandbox.lib.cloudformation_templates import VmBucketCfTemplate
 
 
 @cli.command()
@@ -20,4 +20,4 @@ def setup_vm_bucket_waf(aws_profile: str, allowed_ip: str, log_level: str):
     Command to deploy the VM S3-Bucket Web Application Firewall. Needs to run before deploying the VM Bucket itself.
     """
     set_log_level(log_level)
-    VmBucket.waf(AwsAccess(aws_profile), default_config_object).setup(allowed_ip)
+    VmBucketCfTemplate.waf(AwsAccess(aws_profile), default_config_object).setup(allowed_ip)
