+++ exasol/ds/sandbox/lib/cloudformation_templates/cf_template.py
