+++ exasol/ds/sandbox/lib/cloudformation_templates/vm_bucket.py
@@ -4,44 +4,14 @@ from exasol.ds.sandbox.lib.aws_access.aws_access import AwsAccess
 from exasol.ds.sandbox.lib.config import ConfigObject
 from exasol.ds.sandbox.lib.logging import get_status_logger, LogType
 from exasol.ds.sandbox.lib.asset_id import AssetId
-from exasol.ds.sandbox.lib.aws_access.cloudformation_template import (
+from exasol.ds.sandbox.lib.cloudformation_templates import (
     CfTemplateSpec,
-    CfTemplate,
+    S3BucketCfTemplate,
+    WafCfTemplate,
 )
-from exasol.ds.sandbox.lib.cloudformation.waf import WafCfTemplate
 
 
-LOG = get_status_logger(LogType.S3_BUCKETS)
-
-
-class S3BucketCfTemplate(CfTemplate):
-    """
-    Template for a Cloudformation stack for an AWS S3 bucket.
-    """
-    def __init__(self, aws_access: Optional[AwsAccess], spec: CfTemplateSpec):
-        super().__init__(aws_access, spec)
-
-    def setup(self, waf_acl_arn: str) -> None:
-        super().setup(
-            acl_arn=waf_acl_arn,
-            path_in_bucket=AssetId.BUCKET_PREFIX,
-        )
-        LOG.info(f"Deployed cloudformation stack {self.stack_name}")
-
-    @property
-    def id(self) -> str:
-        return self.stack_output("BucketId")
-
-    @property
-    def url(self) -> str:
-        return self.stack_output("CfDistributionDomainName")
-
-    @property
-    def import_role(self) -> str:
-        return self.stack_output("ExportRoleId")
-
-
-class VmBucket():
+class VmBucketCfTemplate(S3BucketCfTemplate):
     """
     Enables to instantiate templates for cloudformation stacks for an S3
     bucket and the related Web Application Firewall (WAF) dedicated for AI-Lab
@@ -66,9 +36,12 @@ class VmBucket():
         },
     )
 
-    @classmethod
-    def s3_bucket(cls, aws_access: Optional[AwsAccess]) -> S3BucketCfTemplate:
-        return S3BucketCfTemplate(aws_access, cls.S3_BUCKET)
+    def __init__(self, aws_access: Optional[AwsAccess]):
+        super().__init__(aws_access, self.S3_BUCKET)
+
+    def setup(self, config: ConfigObject) -> None:
+        waf_acl_arn = self.waf(self._aws, config).acl_arn
+        super().setup(waf_acl_arn)
 
     @classmethod
     def waf(cls, aws_access: AwsAccess, config: ConfigObject) -> WafCfTemplate:
