+++ test/unit/aws/test_deploy_vm_bucket.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, create_autospec
 import pytest
 
 from exasol.ds.sandbox.lib.aws_access.aws_access import AwsAccess
-from exasol.ds.sandbox.lib.cloudformation.s3_buckets import VmBucket
+from exasol.ds.sandbox.lib.cloudformation_templates import VmBucketCfTemplate
 from test.aws.mock_data import (
     TEST_BUCKET_ID,
     get_waf_cloudformation_mock_data,
@@ -22,8 +22,8 @@ def test_find_bucket_success(test_config):
         get_s3_cloudformation_mock_data() + \
         get_waf_cloudformation_mock_data()
     mock_cast(aws.instantiate_for_region).return_value = aws
-    testee = VmBucket.s3_bucket(aws)
-    testee.setup(VmBucket.waf(aws, test_config).acl_arn)
+    testee = VmBucketCfTemplate(aws)
+    testee.setup(test_config)
     mock_cast(aws.upload_cloudformation_stack).assert_called_once()
     assert TEST_BUCKET_ID == testee.id
 
@@ -37,9 +37,9 @@ def test_vm_bucket_not_deployed(test_config):
     mock_cast(aws.describe_stacks).return_value = get_waf_cloudformation_mock_data()
     mock_cast(aws.instantiate_for_region).return_value = aws
 
-    name = VmBucket.S3_BUCKET.cf_stack_name
-    with pytest.raises(RuntimeError, match=f"Stack {name} not found"):
-        VmBucket.s3_bucket(aws).id
+    testee = VmBucketCfTemplate(aws)
+    with pytest.raises(RuntimeError, match=f"Stack {testee.stack_name} not found"):
+        testee.id
 
 
 def test_waf_not_deployed(test_config):
@@ -50,6 +50,6 @@ def test_waf_not_deployed(test_config):
     aws: Union[AwsAccess, Mock] = create_autospec(AwsAccess, spec_set=True)
     mock_cast(aws.describe_stacks).return_value = list()
 
-    name = VmBucket.S3_BUCKET.cf_stack_name
-    with pytest.raises(RuntimeError, match=f"Stack {name} not found"):
-        VmBucket.s3_bucket(aws).id
+    testee = VmBucketCfTemplate(aws)
+    with pytest.raises(RuntimeError, match=f"Stack {testee.stack_name} not found"):
+        testee.id
