+++ test/unit/aws/test_deploy_vm_bucket_waf.py
@@ -2,7 +2,7 @@ from typing import Union
 from unittest.mock import Mock, create_autospec
 
 from exasol.ds.sandbox.lib.aws_access.aws_access import AwsAccess
-from exasol.ds.sandbox.lib.cloudformation.s3_buckets import VmBucket
+from exasol.ds.sandbox.lib.cloudformation_templates import VmBucketCfTemplate
 from test.aws.templates import TEST_IP, TEST_ACL_ARN
 from test.aws.mock_data import get_waf_cloudformation_mock_data
 from test.mock_cast import mock_cast
@@ -16,7 +16,7 @@ def test_find_acl_arn(test_config):
     aws: Union[AwsAccess, Mock] = create_autospec(AwsAccess, spec_set=True)
     mock_cast(aws.describe_stacks).return_value = get_waf_cloudformation_mock_data()
     mock_cast(aws.instantiate_for_region).return_value = aws
-    testee = VmBucket.waf(aws, config=test_config)
+    testee = VmBucketCfTemplate.waf(aws, config=test_config)
     testee.setup(TEST_IP)
     mock_cast(aws.upload_cloudformation_stack).assert_called_once()
     assert TEST_ACL_ARN == testee.acl_arn
